add new nocopy API
